requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
html2text==2024.2.26
openai==1.3.9
python-dotenv==1.0.0
tavily-python==0.3.3
